fix(checker): bundle symbol flow memo caches
