+++ crates/tsz-checker/src/context/checker_context_lifetimes.toml
@@ -62,7 +62,7 @@ narrowable_identifier_cache = { lifetime = "FileLocalReset", capability = "FlowS
 flow_switch_reference_cache = { lifetime = "FileLocalReset", capability = "FlowSessionState", reason = "checker-local mutable state tied to the active file or traversal; reset at file-session boundary" }
 flow_numeric_atom_cache = { lifetime = "FileLocalReset", capability = "FlowSessionState", reason = "checker-local mutable state tied to the active file or traversal; reset at file-session boundary" }
 flow_reference_match_cache = { lifetime = "FileLocalReset", capability = "FlowSessionState", reason = "checker-local mutable state tied to the active file or traversal; reset at file-session boundary" }
-symbol_last_assignment_pos = { lifetime = "FileLocalReset", capability = "FlowSessionState", reason = "checker-local mutable state tied to the active file or traversal; reset at file-session boundary" }
+symbol_flow_memo = { lifetime = "FileLocalReset", capability = "FlowSessionState", reason = "symbol-stable flow memo tables reused only within the active file session; reset at file-session boundary" }
 symbol_flow_confirmed = { lifetime = "FileLocalReset", capability = "FlowSessionState", reason = "checker-local mutable state tied to the active file or traversal; reset at file-session boundary" }
 call_type_predicates = { lifetime = "FileLocalReset", capability = "FlowSessionState", reason = "checker-local mutable state tied to the active file or traversal; reset at file-session boundary" }
 daa_error_nodes = { lifetime = "DiagnosticsOnly", capability = "DiagnosticState", reason = "diagnostic emission, dedupe, or suppression state; clear at file-session boundary" }
