+++ crates/tsz-checker/src/context/constructors.rs
@@ -113,10 +113,7 @@ impl<'a> CheckerContext<'a> {
             flow_visited: RefCell::new(FxHashSet::default()),
             flow_results: RefCell::new(FxHashMap::with_capacity_and_hasher(64, Default::default())),
             flow_reference_match_cache: RefCell::new(FxHashMap::default()),
-            symbol_last_assignment_pos: RefCell::new(FxHashMap::default()),
-            symbol_nested_closure_assignment: RefCell::new(FxHashMap::default()),
-            symbol_has_non_initializer_assignment: RefCell::new(FxHashMap::default()),
-            symbol_first_identifier_ref: RefCell::new(FxHashMap::default()),
+            symbol_flow_memo: crate::context::SymbolFlowMemoCaches::default(),
             symbol_flow_confirmed: RefCell::new(FxHashMap::default()),
             narrowing_cache: tsz_solver::narrowing::NarrowingCache::new(),
             call_type_predicates: crate::control_flow::CallPredicateMap::default(),
