+++ crates/tsz-checker/src/types/type_node_advanced.rs
@@ -596,11 +596,13 @@ impl<'a, 'ctx> TypeNodeChecker<'a, 'ctx> {
                         &self.ctx.flow_visited,
                         &self.ctx.flow_results,
                     )
-                    .with_symbol_last_assignment_pos(&self.ctx.symbol_last_assignment_pos)
+                    .with_symbol_last_assignment_pos(&self.ctx.symbol_flow_memo.last_assignment_pos)
                     .with_symbol_nested_closure_assignment(
-                        &self.ctx.symbol_nested_closure_assignment,
+                        &self.ctx.symbol_flow_memo.nested_closure_assignment,
                     )
-                    .with_symbol_first_identifier_ref(&self.ctx.symbol_first_identifier_ref)
+                    .with_symbol_first_identifier_ref(
+                        &self.ctx.symbol_flow_memo.first_identifier_ref,
+                    )
                     .with_destructured_bindings(&self.ctx.destructured_bindings);
 
                     let narrowed =
